introduce Twig pattern class
