+++ src/fr/adrienbrault/idea/symfony2plugin/navigation/TwigFoldingBuilder.java
@@ -10,6 +10,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import com.jetbrains.twig.TwigFile;
 import fr.adrienbrault.idea.symfony2plugin.Settings;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
+import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
 import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.routing.Route;
 import fr.adrienbrault.idea.symfony2plugin.routing.RouteHelper;
@@ -56,7 +57,7 @@ public class TwigFoldingBuilder extends FoldingBuilderEx {
     private void attachTemplateFoldingDescriptors(PsiElement psiElement, List<FoldingDescriptor> descriptors) {
         // find path calls in file
         PsiElement[] fileReferences = PsiTreeUtil.collectElements(psiElement, psiElement1 ->
-            TwigHelper.getTemplateFileReferenceTagPattern().accepts(psiElement1) || TwigHelper.getFormThemeFileTag().accepts(psiElement1)
+            TwigPattern.getTemplateFileReferenceTagPattern().accepts(psiElement1) || TwigHelper.getFormThemeFileTag().accepts(psiElement1)
         );
 
         if(fileReferences.length == 0) {
@@ -83,7 +84,7 @@ public class TwigFoldingBuilder extends FoldingBuilderEx {
 
         // find path calls in file
         PsiElement[] psiElements = PsiTreeUtil.collectElements(psiElement, psiElement12 ->
-            TwigHelper.getAutocompletableRoutePattern().accepts(psiElement12)
+            TwigPattern.getAutocompletableRoutePattern().accepts(psiElement12)
         );
 
         if(psiElements.length == 0) {
@@ -122,7 +123,7 @@ public class TwigFoldingBuilder extends FoldingBuilderEx {
     private void attachConstantFoldingDescriptors(PsiElement psiElement, List<FoldingDescriptor> descriptors) {
         // find path calls in file
         PsiElement[] constantReferences = PsiTreeUtil.collectElements(psiElement, psiElement1 ->
-            TwigHelper.getPrintBlockOrTagFunctionPattern("constant").accepts(psiElement1)
+            TwigPattern.getPrintBlockOrTagFunctionPattern("constant").accepts(psiElement1)
         );
 
         if(constantReferences.length == 0) {
