+++ tests/fr/adrienbrault/idea/symfony2plugin/tests/templating/TwigFilterCompletionContributorTest.java
@@ -5,7 +5,7 @@ import com.jetbrains.php.lang.psi.elements.Function;
 import com.jetbrains.php.lang.psi.elements.Method;
 import com.jetbrains.twig.TwigFileType;
 import com.jetbrains.twig.elements.TwigElementTypes;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
 import fr.adrienbrault.idea.symfony2plugin.tests.SymfonyLightCodeInsightFixtureTestCase;
 
 import java.io.File;
@@ -55,7 +55,7 @@ public class TwigFilterCompletionContributorTest extends SymfonyLightCodeInsight
     }
 
     /**
-     * @see TwigHelper#getAfterIsTokenPattern
+     * @see TwigPattern#getAfterIsTokenPattern
      * @see fr.adrienbrault.idea.symfony2plugin.templating.TwigTemplateCompletionContributor.TwigSimpleTestParametersCompletionProvider
      */
     public void testSimpleTestExtension() {
@@ -67,7 +67,7 @@ public class TwigFilterCompletionContributorTest extends SymfonyLightCodeInsight
     }
 
     /**
-     * @see TwigHelper#getAfterOperatorPattern
+     * @see TwigPattern#getAfterOperatorPattern
      * @see fr.adrienbrault.idea.symfony2plugin.templating.TwigTemplateCompletionContributor.TwigSimpleTestParametersCompletionProvider
      */
     public void testOperatorExtension() {
