+++ tests/fr/adrienbrault/idea/symfony2plugin/tests/templating/translation/TwigTranslationCompletionContributorTest.java
@@ -1,7 +1,7 @@
 package fr.adrienbrault.idea.symfony2plugin.tests.templating.translation;
 
 import com.jetbrains.twig.TwigFileType;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
 
 /**
  * @author Daniel Espendiller <daniel@espendiller.net>
@@ -74,7 +74,7 @@ public class TwigTranslationCompletionContributorTest extends TwigTranslationFix
     }
 
     /**
-     * @see TwigHelper#getTransDefaultDomainPattern
+     * @see TwigPattern#getTransDefaultDomainPattern
      * @see fr.adrienbrault.idea.symfony2plugin.templating.TwigTemplateCompletionContributor
      */
     public void testTwigTransDefaultDomainDomainCompletion() {
@@ -84,7 +84,7 @@ public class TwigTranslationCompletionContributorTest extends TwigTranslationFix
     }
 
     /**
-     * @see TwigHelper#getTranslationTokenTagFromPattern
+     * @see TwigPattern#getTranslationTokenTagFromPattern
      */
     public void testTranslationTokenTagFromCompletion() {
         assertCompletionContains(TwigFileType.INSTANCE, "{% trans from \"<caret>\" %}", "interchange");
